Convert to multilingual multisite

+++ wp-content/plugins/email-address-encoder/includes/admin.php
@@ -28,14 +28,34 @@ add_filter( 'plugin_action_links', 'eae_plugin_actions_links', 10, 2 );
 add_action( 'admin_notices', 'eae_page_scanner_notice' );
 
 /**
+ * Register callback to display warnings of incompatible plugins.
+ */
+add_action( 'admin_notices', 'eae_compatibility_warnings' );
+
+/**
+ * Register scripts callback.
+ */
+add_action( 'wp_enqueue_scripts', 'eae_enqueue_scripts' );
+
+/**
  * Register admin scripts callback.
  */
-add_action( 'admin_enqueue_scripts', 'eae_enqueue_script' );
+add_action( 'admin_enqueue_scripts', 'eae_enqueue_admin_scripts' );
 
 /**
  * Register callback to transmit email address to remote server.
  */
 add_action( 'load-settings_page_email-address-encoder', 'eae_transmit_email' );
+
+/**
+ * Register callback to clear page caches.
+ */
+add_action( 'load-options.php', 'eae_clear_caches' );
+
+/**
+ * Register callback that cleans responses for the email detector.
+ */
+add_action( 'wp', 'eae_cleanup_response' );
 
 /**
  * Register AJAX callback for "eae_dismiss_notice" action.
@@ -76,29 +96,10 @@ function eae_register_ui() {
  * @return void
  */
 function eae_register_settings() {
-    register_setting( 'email-address-encoder', 'eae_search_in', array(
-        'type' => 'string',
-        'default' => 'filters',
-        'sanitize_callback' => 'sanitize_text_field',
-    ) );
-
-    register_setting( 'email-address-encoder', 'eae_technique', array(
-        'type' => 'string',
-        'default' => 'entities',
-        'sanitize_callback' => 'sanitize_text_field',
-    ) );
-
-    register_setting( 'email-address-encoder', 'eae_filter_priority', array(
-        'type' => 'integer',
-        'default' => 1000,
-        'sanitize_callback' => 'sanitize_text_field',
-    ) );
-
-    register_setting( 'email-address-encoder', 'eae_notices', array(
-        'type' => 'integer',
-        'default' => 0,
-        'sanitize_callback' => 'intval',
-    ) );
+    register_setting( 'email-address-encoder', 'eae_search_in', 'sanitize_text_field' );
+    register_setting( 'email-address-encoder', 'eae_technique', 'sanitize_text_field' );
+    register_setting( 'email-address-encoder', 'eae_filter_priority', 'absint' );
+    register_setting( 'email-address-encoder', 'eae_notices', 'absint' );
 }
 
 /**
@@ -107,6 +108,14 @@ function eae_register_settings() {
  * @return void
  */
 function eae_uninstall_hook() {
+    if ( ! function_exists( 'get_plugins' ) ) {
+        require_once ABSPATH . 'wp-admin/includes/plugin.php';
+    }
+
+    if ( array_key_exists( 'email-encoder-premium/email-address-encoder.php', get_plugins() ) ) {
+        return;
+    }
+
     delete_option( 'eae_search_in' );
     delete_option( 'eae_technique' );
     delete_option( 'eae_filter_priority' );
@@ -133,7 +142,7 @@ function eae_options_page() {
 }
 
 /**
- * Callback to add "Settings" link to the plugin's action links.
+ * Callback to add links to the plugin's action links.
  *
  * @param array $links
  * @param string $file
@@ -145,13 +154,85 @@ function eae_plugin_actions_links( $links, $file ) {
         return $links;
     }
 
-    $link = sprintf(
-        '<a href="%s">%s</a>',
-        admin_url( 'options-general.php?page=email-address-encoder' ),
-        __( 'Settings', 'email-address-encoder' )
-    );
+    return array_merge( array(
+        sprintf(
+            '<a target="_blank" rel="noopener" href="https://encoder.till.im/guide?utm_source=wp-plugin&amp;utm_medium=action-link">%s</a>',
+            __( 'FAQ', 'email-address-encoder' )
+        ),
+        sprintf(
+            '<a target="_blank" rel="noopener" href="https://encoder.till.im/download?utm_source=wp-plugin&amp;utm_medium=action-link">%s</a>',
+            __( 'Premium', 'email-address-encoder' )
+        ),
+        sprintf(
+            '<a href="%s">%s</a>',
+            admin_url( 'options-general.php?page=email-address-encoder' ),
+            __( 'Settings', 'email-address-encoder' )
+        ),
+    ), $links );
+}
 
-    return array_merge( array( $link ), $links );
+/**
+ * Callback to load email detector script.
+ *
+ * @return void
+ */
+function eae_enqueue_scripts() {
+    if ( ! is_admin_bar_showing() ) {
+        return;
+    }
+
+    if ( is_preview() ) {
+        return;
+    }
+
+    if ( ! current_user_can( 'manage_options' ) ) {
+        return;
+    }
+
+    if ( defined( 'EAE_DISABLE_NOTICES' ) && EAE_DISABLE_NOTICES ) {
+        return;
+    }
+
+    if ( get_option( 'eae_notices', '0' ) == '1' ) {
+        return;
+    }
+
+    add_action( 'wp_footer', 'eae_adminbar_styles' );
+
+    $version = get_file_data( __DIR__ . '/../email-address-encoder.php', array( 'Version' => 'Version' ) );
+
+    wp_enqueue_script( 'email-detector', plugins_url( 'email-detector.js', __FILE__ ), null, $version[ 'Version' ], true );
+
+    wp_localize_script( 'email-detector', 'eae_detector', array(
+        'one_email' => __( '1 Unprotected Email', 'email-address-encoder' ),
+        'many_emails' => __( '{number} Unprotected Emails', 'email-address-encoder' ),
+    ) );
+
+    add_filter( 'script_loader_tag', 'eae_defer_script', 10, 2 );
+}
+
+/**
+ * Callback to load email detector script.
+ *
+ * @return void
+ */
+function eae_adminbar_styles() {
+    $styles = '#wp-admin-bar-eae > .ab-item:before { content: "\f534"; top: 2px; }';
+
+    echo "\n<style type=\"text/css\">{$styles}</style>\n";
+}
+
+/**
+ * Defer email detector script.
+ *
+ * @return void
+ */
+function eae_defer_script( $tag, $handle ) {
+    if ( $handle === 'email-detector' ) {
+        return str_replace( ' src', ' defer="defer" src', $tag );
+    }
+
+    return $tag;
 }
 
 /**
@@ -159,17 +240,21 @@ function eae_plugin_actions_links( $links, $file ) {
  *
  * @return void
  */
-function eae_enqueue_script() {
+function eae_enqueue_admin_scripts() {
     $screen = get_current_screen();
 
-    if ( ! isset( $screen->id ) || $screen->id !== 'dashboard' ) {
+    if ( ! isset( $screen->id ) ) {
+        return;
+    }
+
+    if ( ! in_array( $screen->id, array( 'dashboard', 'edit-page' ) ) ) {
         return;
     }
 
     wp_enqueue_script(
         'dismissible-notices',
         plugins_url( 'dismiss-notice.js', __FILE__ ),
-        array( 'jquery', 'common' )
+        array( 'jquery' )
     );
 }
 
@@ -201,7 +286,7 @@ function eae_page_scanner_notice() {
         return;
     }
 
-    if ( $screen->id !== 'dashboard' && $screen->id !== 'edit-page' ) {
+    if ( ! in_array( $screen->id, array( 'dashboard', 'edit-page' ) ) ) {
         return;
     }
 
@@ -213,11 +298,11 @@ function eae_page_scanner_notice() {
         return;
     }
 
-    if ( get_option( 'eae_notices', '0' ) === '1' ) {
+    if ( get_option( 'eae_notices', '0' ) == '1' ) {
         return;
     }
 
-    if ( get_user_meta( get_current_user_id(), 'eae_dismissed_automatic_warnings_notice', true ) === '1' ) {
+    if ( get_user_meta( get_current_user_id(), 'eae_dismissed_automatic_warnings_notice', true ) == '1' ) {
         return;
     }
 
@@ -245,6 +330,22 @@ function eae_transmit_email() {
         return;
     }
 
+    $host = parse_url( get_home_url(), PHP_URL_HOST );
+
+    if (
+        $host === 'localhost' ||
+        filter_var( $host, FILTER_VALIDATE_IP ) ||
+        preg_match( '/\.(dev|test|local)$/', $host ) ||
+        preg_match( '/^(dev|test|staging)\./', $host )
+    ) {
+        return add_settings_error(
+            'eae_notify_email',
+            'invalid',
+            sprintf( __( 'Sorry, "%s" doesn’t appear to be a production domain.', 'email-address-encoder' ), $host ),
+            'error'
+        );
+    }
+
     check_admin_referer( 'subscribe' );
 
     $response = wp_remote_post( 'https://encoder.till.im/api/subscribe', array(
@@ -258,14 +359,12 @@ function eae_transmit_email() {
     ) );
 
     if ( is_wp_error( $response ) || $response[ 'response' ][ 'code' ] !== 200 ) {
-        add_settings_error(
+        return add_settings_error(
             'eae_notify_email',
             'invalid',
             __( 'Whoops, something went wrong. Please try again.', 'email-address-encoder' ),
             'error'
         );
-
-        return;
     }
 
     add_settings_error(
@@ -274,4 +373,121 @@ function eae_transmit_email() {
         __( 'You’ll receive a notification should your site contain unprotected email addresses.', 'email-address-encoder' ),
         'updated'
     );
+}
+
+/**
+ * Clear page caches caches.
+ *
+ * @return void
+ */
+function eae_clear_caches() {
+    if (
+        empty( $_POST ) ||
+        ! isset( $_POST[ 'option_page' ] ) ||
+        $_POST[ 'option_page' ] !== 'email-address-encoder'
+    ) {
+        return;
+    }
+
+    // W3 Total Cache
+    if ( function_exists( 'w3tc_flush_all' ) ) {
+        w3tc_flush_all();
+    }
+
+    // WP Rocket
+    if ( function_exists( 'rocket_clean_domain' ) ) {
+        rocket_clean_domain();
+    }
+
+    // WP Super Cache
+    if ( function_exists( 'wp_cache_clear_cache' ) ) {
+        wp_cache_clear_cache();
+    }
+
+    // JCH Optimize
+    if ( class_exists( 'JchPlatformCache' ) && method_exists( 'JchPlatformCache', 'deleteCache' ) ) {
+        JchPlatformCache::deleteCache( true );
+    }
+
+    // LiteSpeed Cache
+    if ( class_exists( 'LiteSpeed_Cache_API' ) && method_exists( 'LiteSpeed_Cache_API', 'purge_all' ) ) {
+        LiteSpeed_Cache_API::purge_all();
+    }
+
+    // Cachify
+    if ( class_exists( 'Cachify' ) && method_exists( 'Cachify', 'flush_total_cache' ) ) {
+        Cachify::flush_total_cache( true );
+    }
+}
+
+/**
+ * Remove output that might contain email addresses that
+ * would lead to false-positive matches.
+ *
+ * @return void
+ */
+function eae_cleanup_response() {
+    if ( ! isset( $_SERVER[ 'HTTP_X_EMAIL_DETECTOR' ] ) || $_SERVER[ 'HTTP_X_EMAIL_DETECTOR' ] !== 'true' ) {
+        return;
+    }
+
+    // Disable Admin Bar
+    add_filter( 'show_admin_bar', '__return_false' );
+
+    // Disable Debug Bar
+    add_filter( 'debug_bar_enable', '__return_false' );
+
+    // Disable Query Monitor
+    add_filter( 'user_has_cap', function ( $caps ) {
+        $caps[ 'view_query_monitor' ] = false;
+
+        return $caps;
+    } );
+}
+
+/**
+ * Display warnings when incompatible plugins are detected.
+ *
+ * @return void
+ */
+function eae_compatibility_warnings() {
+    $screen = get_current_screen();
+
+    if ( ! isset( $screen->id ) ) {
+        return;
+    }
+
+    $screens = array(
+        'dashboard',
+        'plugins',
+        'edit-page',
+        'settings_page_email-address-encoder'
+    );
+
+    if ( ! in_array( $screen->id, $screens ) ) {
+        return;
+    }
+
+    if ( ! current_user_can( 'manage_options' ) ) {
+        return;
+    }
+
+    $plugins = apply_filters( 'active_plugins', get_option( 'active_plugins' ) );
+
+    foreach ( $plugins as $plugin ) {
+        if ( strpos( $plugin, 'ginger/' ) === 0 ) {
+            $gingerCookieInstalled = true;
+        }
+    }
+
+    if ( isset( $gingerCookieInstalled ) ) {
+        printf(
+            '<div class="notice notice-error"><p><strong>%s</strong> %s</p></div>',
+            __( 'Incompatible plugin detected!', 'email-address-encoder' ),
+            sprintf(
+                __( 'The "Ginger – EU Cookie Law" plugin decodes all HTML entities and thus prevents the Email Address Encoder from working. Please use a different cookie banner plugin, or use the full-page scanner technique of the <a href="%s">Premium version</a>.', 'email-address-encoder' ),
+                admin_url( 'options-general.php?page=email-address-encoder' )
+            )
+        );
+    }
 }
